+++ pkg/utils/eden.go
@@ -87,7 +87,7 @@ func StatusAdam() (status string, err error) {
 //StartEServer function run eserver to serve images
 func StartEServer(commandPath string, serverPort string, imageDist string, logFile string, pidFile string) (err error) {
 	commandArgsString := fmt.Sprintf("server -p %s -d %s -v %s", serverPort, imageDist, log.GetLevel())
-	log.Debugf("StartEServer run: %s %s", commandPath, commandArgsString)
+	log.Infof("StartEServer run: %s %s", commandPath, commandArgsString)
 	return RunCommandNohup(commandPath, logFile, pidFile, strings.Fields(commandArgsString)...)
 }
 
@@ -103,9 +103,9 @@ func StatusEServer(pidFile string) (status string, err error) {
 
 //StartEVEQemu function run EVE in qemu
 func StartEVEQemu(commandPath string, qemuARCH string, qemuOS string, eveImageFile string, qemuSMBIOSSerial string, qemuAccel bool, qemuConfigFilestring, logFile string, pidFile string) (err error) {
-	commandArgsString := fmt.Sprintf("eve qemurun --qemu-config=%s --eve-serial=%s --eve-accel=%t --eve-arch=%s --eve-os=%s --eve-log=%s --eve-pid=%s --image-file=%s -v %s",
+	commandArgsString := fmt.Sprintf("eve start --qemu-config=%s --eve-serial=%s --eve-accel=%t --eve-arch=%s --eve-os=%s --eve-log=%s --eve-pid=%s --image-file=%s -v %s",
 		qemuConfigFilestring, qemuSMBIOSSerial, qemuAccel, qemuARCH, qemuOS, logFile, pidFile, eveImageFile, log.GetLevel())
-	log.Debugf("StartEVEQemu run: %s %s", commandPath, commandArgsString)
+	log.Infof("StartEVEQemu run: %s %s", commandPath, commandArgsString)
 	return RunCommandWithLogAndWait(commandPath, logLevelToPrint, strings.Fields(commandArgsString)...)
 }
 
@@ -127,7 +127,7 @@ func GenerateEveCerts(commandPath string, certsDir string, domain string, ip str
 		}
 	}
 	commandArgsString := fmt.Sprintf("certs --certs-dist=%s --domain=%s --ip=%s --uuid=%s -v %s", certsDir, domain, ip, uuid, log.GetLevel())
-	log.Debugf("GenerateEveCerts run: %s %s", commandPath, commandArgsString)
+	log.Infof("GenerateEveCerts run: %s %s", commandPath, commandArgsString)
 	return RunCommandWithLogAndWait(commandPath, logLevelToPrint, strings.Fields(commandArgsString)...)
 }
 
@@ -188,7 +188,7 @@ func CloneFromGit(dist string, gitRepo string, tag string) (err error) {
 		tag = "master"
 	}
 	commandArgsString := fmt.Sprintf("clone --branch %s --single-branch %s %s", tag, gitRepo, dist)
-	log.Debugf("CloneFromGit run: %s %s", "git", commandArgsString)
+	log.Infof("CloneFromGit run: %s %s", "git", commandArgsString)
 	return RunCommandWithLogAndWait("git", logLevelToPrint, strings.Fields(commandArgsString)...)
 }
 
@@ -202,7 +202,7 @@ func DownloadEveFormDocker(commandPath string, dist string, arch string, tag str
 	}
 	commandArgsString := fmt.Sprintf("eve download --eve-tag=%s --eve-arch=%s -d %s --baseos=%t -v %s",
 		tag, arch, filepath.Join(dist, "dist", arch), baseOs, log.GetLevel())
-	log.Debugf("DownloadEveFormDocker run: %s %s", commandPath, commandArgsString)
+	log.Infof("DownloadEveFormDocker run: %s %s", commandPath, commandArgsString)
 	return RunCommandWithLogAndWait(commandPath, logLevelToPrint, strings.Fields(commandArgsString)...)
 }
 
@@ -218,7 +218,7 @@ func ChangeConfigPartAndRootFs(commandPath string, distEve string, distAdam stri
 	}
 	commandArgsString := fmt.Sprintf("eve confchanger --image-file=%s --config-part=%s --hv=%s -v %s",
 		imagePath, configPath, hv, log.GetLevel())
-	log.Debugf("ChangeConfigPartAndRootFs run: %s %s", commandPath, commandArgsString)
+	log.Infof("ChangeConfigPartAndRootFs run: %s %s", commandPath, commandArgsString)
 	return RunCommandWithLogAndWait(commandPath, logLevelToPrint, strings.Fields(commandArgsString)...)
 }
 
@@ -236,23 +236,23 @@ func MakeEveInRepo(distEve string, distAdam string, arch string, hv string, root
 	if rootFSOnly {
 		commandArgsString := fmt.Sprintf("-C %s HV=%s CONF_DIR=%s rootfs",
 			distEve, hv, configPath)
-		log.Debugf("MakeEveInRepo run: %s %s", "make", commandArgsString)
+		log.Infof("MakeEveInRepo run: %s %s", "make", commandArgsString)
 		err = RunCommandWithLogAndWait("make", logLevelToPrint, strings.Fields(commandArgsString)...)
 	} else {
 		commandArgsString := fmt.Sprintf("-C %s HV=%s CONF_DIR=%s IMG_FORMAT=qcow2 live",
 			distEve, hv, configPath)
-		log.Debugf("MakeEveInRepo run: %s %s", "make", commandArgsString)
+		log.Infof("MakeEveInRepo run: %s %s", "make", commandArgsString)
 		err = RunCommandWithLogAndWait("make", logLevelToPrint, strings.Fields(commandArgsString)...)
 		biosPath := filepath.Join(distEve, "dist", arch, "OVMF.fd")
 		commandArgsString = fmt.Sprintf("-C %s HV=%s %s",
 			distEve, hv, biosPath)
-		log.Debugf("MakeEveInRepo run: %s %s", "make", commandArgsString)
+		log.Infof("MakeEveInRepo run: %s %s", "make", commandArgsString)
 		err = RunCommandWithLogAndWait("make", logLevelToPrint, strings.Fields(commandArgsString)...)
 		if arch == "arm64" {
 			dtbPath := filepath.Join(distEve, "dist", "eve.dtb")
 			commandArgsString = fmt.Sprintf("-C %s HV=%s %s",
 				distEve, hv, dtbPath)
-			log.Debugf("MakeEveInRepo run: %s %s", "make", commandArgsString)
+			log.Infof("MakeEveInRepo run: %s %s", "make", commandArgsString)
 			err = RunCommandWithLogAndWait("make", logLevelToPrint, strings.Fields(commandArgsString)...)
 		}
 	}
@@ -270,13 +270,13 @@ func BuildVM(linuxKitPath string, imageConfig string, distImage string) (err err
 	imageConfigTmp := filepath.Join(distImageDir, fmt.Sprintf("%s-bios.img", fileNameWithoutExtension(filepath.Base(distImage))))
 	commandArgsString := fmt.Sprintf("build -format raw-bios -dir %s %s",
 		distImageDir, imageConfig)
-	log.Debugf("BuildVM run: %s %s", linuxKitPath, commandArgsString)
+	log.Infof("BuildVM run: %s %s", linuxKitPath, commandArgsString)
 	if err = RunCommandWithLogAndWait(linuxKitPath, logLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
 		return fmt.Errorf("error in linuxkit: %s", err)
 	}
 	commandArgsString = fmt.Sprintf("convert -c -f raw -O qcow2 %s %s",
 		imageConfigTmp, distImage)
-	log.Debugf("BuildVM run: %s %s", "qemu-img", commandArgsString)
+	log.Infof("BuildVM run: %s %s", "qemu-img", commandArgsString)
 	if err = RunCommandWithLogAndWait("qemu-img", logLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
 		return fmt.Errorf("error in qemu-img: %s", err)
 	}
@@ -285,16 +285,10 @@ func BuildVM(linuxKitPath string, imageConfig string, distImage string) (err err
 
 //PrepareQEMUConfig create config file for QEMU
 func PrepareQEMUConfig(commandPath string, qemuConfigFile string, firmwareFile []string, configPart string, dtbPath string, eveHostFWD string) (err error) {
-	if _, err := os.Stat(configPart); os.IsNotExist(err) {
-		return fmt.Errorf("cannot find configPart %s: %s", configPart, err)
-	}
 	firmwares := strings.Join(firmwareFile, ",")
-	commandArgsString := fmt.Sprintf("eve qemuconf --qemu-config=%s --eve-firmware=%s --config-part=%s --eve-hostfwd=\"%s\" -v %s",
-		qemuConfigFile, firmwares, configPart, eveHostFWD, log.GetLevel())
-	if _, err := os.Stat(dtbPath); !os.IsNotExist(err) {
-		commandArgsString = fmt.Sprintf("%s --dtb-part=%s", commandArgsString, dtbPath)
-	}
-	log.Debugf("PrepareQEMUConfig run: %s %s", commandPath, commandArgsString)
+	commandArgsString := fmt.Sprintf("eve qemuconf --qemu-config=%s --eve-firmware=%s --config-part=%s --eve-hostfwd=\"%s\" --dtb-part=%s -v %s",
+		qemuConfigFile, firmwares, configPart, eveHostFWD, dtbPath, log.GetLevel())
+	log.Infof("PrepareQEMUConfig run: %s %s", commandPath, commandArgsString)
 	if err = RunCommandWithLogAndWait(commandPath, logLevelToPrint, strings.Fields(commandArgsString)...); err != nil {
 		return fmt.Errorf("error in qemuconf: %s", err)
 	}
@@ -302,10 +296,10 @@ func PrepareQEMUConfig(commandPath string, qemuConfigFile string, firmwareFile [
 }
 
 //CleanEden teardown Eden and cleanup
-func CleanEden(commandPath, eveDist, eveBaseDist, adamDist, certsDist, imagesDist, binDir, qemuConfFile, eserverPID, evePID string) (err error) {
+func CleanEden(commandPath, eveDist, eveBaseDist, adamDist, certsDist, imagesDist, binDir, eserverPID, evePID string) (err error) {
 	commandArgsString := fmt.Sprintf("stop --eserver-pid=%s --eve-pid=%s --adam-rm=true",
 		eserverPID, evePID)
-	log.Debugf("CleanEden run: %s %s", commandPath, commandArgsString)
+	log.Infof("CleanEden run: %s %s", commandPath, commandArgsString)
 	_, _, err = RunCommandAndWait(commandPath, strings.Fields(commandArgsString)...)
 	if err != nil {
 		return fmt.Errorf("error in eden stop: %s", err)
@@ -328,11 +322,6 @@ func CleanEden(commandPath, eveDist, eveBaseDist, adamDist, certsDist, imagesDis
 	if _, err = os.Stat(imagesDist); !os.IsNotExist(err) {
 		if err = os.RemoveAll(imagesDist); err != nil {
 			return fmt.Errorf("error in %s delete: %s", imagesDist, err)
-		}
-	}
-	if _, err = os.Stat(qemuConfFile); !os.IsNotExist(err) {
-		if err = os.RemoveAll(qemuConfFile); err != nil {
-			return fmt.Errorf("error in %s delete: %s", qemuConfFile, err)
 		}
 	}
 	if _, err = os.Stat(adamDist); !os.IsNotExist(err) {
